change to float functions

+++ include/laml/constants.hpp
@@ -7,9 +7,9 @@ namespace rh {
     namespace laml {
 
         namespace constants {
-            const f64 pi = 3.14159265358979323846;
-            const f64 deg2rad = pi / 180.0;
-            const f64 rad2deg = 180.0 / pi;
+            const f32 pi = 3.14159265358979323846f;
+            const f32 deg2rad = pi / 180.0f;
+            const f32 rad2deg = 180.0f / pi;
         }
     }
 }
